refactor(tui): split app.rs into app/, executor/ modules (SRP, closes #151)
Extract six single-responsibility modules from the 733-line tui/src/app.rs:

executor/classifier.rs  — LogKind + LogLine::classify() (pure, zero I/O)
executor/tailer.rs      — CurrentLog struct + read_new() / drain_log()
executor/mod.rs         — ExecutorState with poll() split into five focused
                          sub-methods: check_process, check_log_completed,
                          find_new_log, tail_current, update_scroll
app/state.rs            — TaskStatus, Task, parse_task_brief, refresh_tasks,
                          select_next/prev, task_counts
app/input.rs            — handle_key, mode-specific dispatch, retry_selected,
                          start_executor
app/mod.rs              — App struct, Mode enum, new(), on_tick() (< 100 lines)

Also fixes the broken ui/ files that referenced the old View/task::Status API:

ui/mod.rs    — dispatch on Mode (TaskList / Executor)
ui/list.rs   — rewritten; uses ListState stateful widget, TaskStatus colours
ui/detail.rs — repurposed as executor log view with colour-coded LogKind lines

All &PathBuf function parameters changed to &Path to pass clippy::ptr_arg.
Tests from the old app.rs are preserved in classifier.rs and state.rs.
